Function to check the data passed to slmfit
The idea is to use this function inside slmfit in line 87 add:

data<-sptotal_datacheck(data,xcoordcol,ycoordcol)

It will handle passing a string with a shapfile path to slmfit. If this
idea is OK, then we need to ad rgdal to the dependencies.

To discuss: Give the user the option to 
project data in lat-long  to some other system. He would need to know
the projection to use and that can be anything. 

We could let the user pass a string that can be passed to CRS, that
option is OK, but for people that do not know much about coordinate
systems that can be tricky, maybe better to ask the user to do the data
preparation beforehand and not assume too much.